pyproject.toml
[project]
name = "pss-fleet-data-importer"
version = "0.1.0"
description = "A tool to import PSS Fleet Data from google drive to the PSS Fleet Data API."
authors = [
    { name = "The worst.", email = "theworstpss@gmail.com" }
]
dependencies = [
    "alembic>=1.13.2",
    "asyncpg>=0.29.0",
    "cancel-token>=0.1.6",
    "pssapi>=0.5.1",
    "pss-fleet-data-client>=0.3.1",
    "psycopg2-binary>=2.9.9",
    "PyDrive2>=1.20.0",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.20",
    "stopit>=1.1.2",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "aiosqlite>=0.20.0",
    "autoflake>=2.3.1",
    "black>=24.4.2",
    "coverage>=7.6.0",
    "flake8>=7.1.0",
    "flake8-black>=0.3.6",
    "flake8-bugbear>=24.4.26",
    "isort>=5.13.2",
    "pre-commit>=3.8.0",
    "pytest>=7.2.1",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "pytest-env>=1.1.3",
    "pytest_httpx>=0.30.0",
    "pytest-mock>=3.14.0",
    "pytest-vcr>=1.0.2",
    "vulture>=2.11",
]

[tool.autoflake]
expand-star-imports = true
ignore-init-module-imports = true
in-place = true
recursive = true
remove-all-unused-imports = true

[tool.black]
line-length = 150
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
src_paths = ["src", "tests"]
profile = "black"
line_length = 150
lines_after_imports = 2

[tool.vulture]
min_confidence = 100
paths = ["./src/app"]
sort_by_size = true
